refactor(impl): comment and refactor on events
add comments to event processing and also extract functions for receiving requests to a new file
